Fixed PR comments
- Moved ToggleArrayElement into sharedUtilities
- changed the MostActiveUser's filter criteria into variables to be more readable and explicit
- added img alt for the profile pictures in the MostActiveUser's section
